Splitting DAsm.cpp into files for each class.
