+++ DAsm.h
@@ -15,150 +15,56 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 #include <cstdint>
 #include <list>
 #include <string>
+#include <functional>
 
 namespace DAsm{
+    typedef uint16_t word;
+    typedef uint8_t opcode; //NB opcodes only actually need 5 bits
 
-typedef uint16_t word;
-typedef uint8_t opcode; //NB opcodes only actually need 5 bits
+    const unsigned long DCPU_RAM_SIZE = 65536;
 
-const unsigned long DCPU_RAM_SIZE = 65536;
+    const unsigned int MAX_RECURSION = 20;
 
-const unsigned int MAX_RECURSION = 20;
+    struct  str_opcode{
+        std::string  str;
+        opcode  value;
+        bool arguments;
+        str_opcode(std::string nStr, opcode nValue, bool nArguments = true){str=nStr; value=nValue; arguments=nArguments;}
+    };
+    //Store words to be replaced with expression values
+    struct  expression_target{
+        std::string expression;
+        word*   target;
+        expression_target(std::string nExpression, word* nTarget){expression=nExpression; target=nTarget;}
+    };
 
-struct  str_opcode{
-    std::string  str;
-    opcode  value;
-    bool arguments;
-    str_opcode(std::string nStr, opcode nValue, bool nArguments = true){str=nStr; value=nValue; arguments=nArguments;}
-};
-//Store words to be replaced with expression values
-struct  expression_target{
-    std::string expression;
-    word*   target;
-    expression_target(std::string nExpression, word* nTarget){expression=nExpression; target=nTarget;}
-};
+    //Store values of labels
+    //Stores int-width values for extra-long defines
+    struct label_value{
+        std::string label;
+        int    value;
+        label_value(std::string nLabel, int nValue){label=nLabel; value=nValue;}
+    };
 
-//Store values of labels
-//Stores int-width values for extra-long defines
-struct label_value{
-    std::string label;
-    int    value;
-    label_value(std::string nLabel, int nValue){label=nLabel; value=nValue;}
-};
+    struct macro{
+        std::string keyword;
+        std::string format;
+    };
 
-struct macro{
-    std::string keyword;
-    std::string format;
-};
+    std::string trimWS(std::string& s);
 
-class   Program;
+    //std::stoi is far too lenient wrt numbers with other stuff at the end
+    int     getNumber(std::string str, bool& number);
 
-class   Instruction{
-    public:
+    //Function to split string based on regex and put in list. Default predicate is just non-emptiness
+    void    splitString(std::string& source, std::string splitRegex, std::list<std::string>& stringList,
+                        std::function<bool(std::string)> predicate = [](std::string str){return (str.size() > 0);});
 
+    //Search/replace by regex
+    std::string replaceString(std::string source, std::string searchRegex, std::string replaceStr);
 
-                                            Instruction();
-                                            Instruction(std::string source, unsigned int recursion_counter = 0);
-        virtual                             ~Instruction();
-
-        void                                LoadSource(std::string source, unsigned int recursion_counter = 0);
-
-        void                                Fill(word value, unsigned int amount);
-
-        unsigned int                        GetLength();
-
-        word                                ParseArg(std::string source, bool isA = false);
-
-        opcode                              GetOpcode(std::string str);
-        opcode                              GetSpecialOpcode(std::string str);
-
-        void                                Error(std::string nError);
-
-        std::list<word>                     mWords;
-
-        static std::list<str_opcode>        mOpcodes;
-        static std::list<str_opcode>        mSpecialOpcodes;
-        static std::list<str_opcode>        mReg;
-
-        static Program*                     mProgram;
-
-};
-
-class   ProgramChunk{
-    public:
-
-
-                                            ProgramChunk();
-        virtual                             ~ProgramChunk();
-
-        bool                                mHasTargetPos;
-        word                                mTargetPos;
-
-        unsigned int                        GetLength();
-
-        void                                SetLabels();
-
-
-        std::list<label_value>              mLabelValues;
-        std::list<Instruction>              mInstructions;
-};
-
-class   Program{
-    public:
-                                            Program();
-        virtual                             ~Program();
-
-        bool                                LoadSource(std::string source);
-
-        void                                AddExpressionTarget(std::string nExpression,word* nTarget);
-
-        void                                AddDefineOnlyTarget(std::string nExpression,word* nTarget);
-
-        void                                AddIncrementTarget(std::string nExpression,word* nTarget);
-
-        void                                AddLabelValue(std::string nLabel, word nValue);
-        void                                AddDefine(std::string nLabel, int nValue);
-        std::string                         GlobalizeLabels(std::string nExpression);
-
-        unsigned int                        GetLength();
-        unsigned int                        mLength;
-
-        std::string                         ToHex(std::string seperator = " 0x");
-
-        int                                 Evaluate(std::string expression, bool* errorFlag=nullptr);
-
-        bool                                IsMacro(std::string keyword);
-        std::string                         DoMacro(std::string keyword, std::list<std::string> args);
-        void                                AddMacro(std::string nMacro);
-
-        std::list<macro>                    mMacros;
-
-        word*                               ToBlock(unsigned long maxSize = DCPU_RAM_SIZE);
-        word*                               ToBlock(unsigned long maxSize, unsigned long& programSize);
-
-        void                                Error(std::string nError);
-        std::list<std::string>              mErrors;
-
-        std::list<ProgramChunk>             mChunks;
-        std::list<ProgramChunk*>            mOrdered;
-
-        std::string                         mGlobalLabel;
-
-        std::list<label_value>              mDefineValues;
-
-        std::list<expression_target>        mDefineOnlyTargets;
-
-        std::list<expression_target>        mExpressionTargets;
-        std::list<expression_target>        mIncrementTargets;
-
-        std::list<Instruction>*             mInstructions;
-        ProgramChunk*                       mCurChunk;
-
-        bool                                mStrictDefineCommas;
-        bool                                mIgnoreLabelCase; //When true, all labels are converted to uppercase
-        bool                                mArrangeChunks;//When true, .orgs control actual position of code in output
-};
-
+    //Escape a string so it can be matched as a regex
+    std::string regexEscape(std::string source);
 }
 
 #endif // DASM_H
